OpenNL: modify SuperLU to use doubles rather than floats, for better precision.
This helps to improve the accuracy of UV unwrapping and laplacian deform for
high poly meshes, which could get warped quite badly. It's not much slower,
doubles are pretty fast on modern CPUs, but it does double memory usage. This
seems acceptable as otherwise high poly meshes would not work correctly anyway.

Fixes T39004.

+++ intern/opennl/superlu/scopy_to_ucol.c
@@ -33,7 +33,7 @@ scopy_to_ucol(
 	      int        *segrep,  /* in */
 	      int        *repfnz,  /* in */
 	      int        *perm_r,  /* in */
-	      float     *dense,   /* modified - reset to zero on return */
+	      double     *dense,   /* modified - reset to zero on return */
 	      GlobalLU_t *Glu      /* modified */
 	      )
 {
@@ -47,11 +47,11 @@ scopy_to_ucol(
     int new_next, mem_error;
     int       *xsup, *supno;
     int       *lsub, *xlsub;
-    float    *ucol;
+    double    *ucol;
     int       *usub, *xusub;
     int       nzumax;
 
-    float zero = 0.0;
+    double zero = 0.0;
 
     xsup    = Glu->xsup;
     supno   = Glu->supno;
